修改目标方法调用方式，增加并发处理
缓存接口时间设置为int型，以适应jedis；增加代理方法调用时的同步代码块，防止并发异常

+++ src/main/java/com/oliver/autocache/annotation/Cache.java
@@ -31,7 +31,7 @@ public @interface Cache {
      *
      * @return
      */
-    long time() default CacheTime.TIME_MIDDLE;
+    int time() default CacheTime.TIME_MIDDLE;
 
 
 }
